uuid: Move the uuid package into its own top level directory
There was no real reason to have this package embedded inside the ssntp
package and having it in there confuses dep prune making the package
harder to vendor.
