+++ .travis.yml
@@ -68,7 +68,7 @@ script:
    - sudo mkdir -p /var/lib/ciao/data/controller/workloads
    - sudo chmod 0777 -R /var/lib/ciao
    - if [ -z $LINT_ONLY ]; then test-cases $TEST_CASES_FLAGS -v -timeout 9 -coverprofile /tmp/cover.out -short github.com/ciao-project/ciao/ciao-controller/...; fi
-   - if [ -z $LINT_ONLY ]; then test-cases $TEST_CASES_FLAGS -v -timeout 9 -coverprofile /tmp/cover.out -append-profile -short github.com/ciao-project/ciao/ciao-launcher github.com/ciao-project/ciao/ciao-scheduler github.com/ciao-project/ciao/payloads github.com/ciao-project/ciao/configuration github.com/ciao-project/ciao/testutil github.com/ciao-project/ciao/ssntp/uuid github.com/ciao-project/ciao/qemu github.com/ciao-project/ciao/openstack/... github.com/ciao-project/ciao/bat  github.com/ciao-project/ciao/ciao-image/... github.com/ciao-project/ciao/database/... github.com/ciao-project/ciao/ciao-storage/... github.com/ciao-project/ciao/deviceinfo github.com/ciao-project/ciao/osprepare; fi
+   - if [ -z $LINT_ONLY ]; then test-cases $TEST_CASES_FLAGS -v -timeout 9 -coverprofile /tmp/cover.out -append-profile -short github.com/ciao-project/ciao/ciao-launcher github.com/ciao-project/ciao/ciao-scheduler github.com/ciao-project/ciao/payloads github.com/ciao-project/ciao/configuration github.com/ciao-project/ciao/testutil github.com/ciao-project/ciao/uuid github.com/ciao-project/ciao/qemu github.com/ciao-project/ciao/openstack/... github.com/ciao-project/ciao/bat  github.com/ciao-project/ciao/ciao-image/... github.com/ciao-project/ciao/database/... github.com/ciao-project/ciao/ciao-storage/... github.com/ciao-project/ciao/deviceinfo github.com/ciao-project/ciao/osprepare; fi
    - if [ -z $LINT_ONLY ]; then export GOROOT=`go env GOROOT` && sudo -E PATH=$PATH:$GOROOT/bin $GOPATH/bin/test-cases $TEST_CASES_FLAGS -v -timeout 9 -coverprofile /tmp/cover.out -append-profile github.com/ciao-project/ciao/ssntp; fi
    - if [ -z $LINT_ONLY ]; then export GOROOT=`go env GOROOT` && export SNNET_ENV=198.51.100.0/24 && sudo -E PATH=$PATH:$GOROOT/bin $GOPATH/bin/test-cases -race -v -timeout 9 -short -tags travis -coverprofile /tmp/cover.out -append-profile github.com/ciao-project/ciao/networking/libsnnet; fi
