+++ _release/bat/image_bat/image_bat_test.go
@@ -23,7 +23,7 @@ import (
 	"time"
 
 	"github.com/ciao-project/ciao/bat"
-	"github.com/ciao-project/ciao/ssntp/uuid"
+	"github.com/ciao-project/ciao/uuid"
 )
 
 const standardTimeout = time.Second * 300
