+++ ciao-controller/external_ip.go
@@ -19,7 +19,7 @@ import (
 
 	"github.com/ciao-project/ciao/ciao-controller/types"
 	"github.com/ciao-project/ciao/payloads"
-	"github.com/ciao-project/ciao/ssntp/uuid"
+	"github.com/ciao-project/ciao/uuid"
 )
 
 func (c *controller) makePoolLinks(pool *types.Pool) {
