+++ ciao-controller/image.go
@@ -25,7 +25,7 @@ import (
 	"github.com/ciao-project/ciao/ciao-storage"
 	"github.com/ciao-project/ciao/database"
 	"github.com/ciao-project/ciao/payloads"
-	"github.com/ciao-project/ciao/ssntp/uuid"
+	"github.com/ciao-project/ciao/uuid"
 	"github.com/golang/glog"
 	"github.com/pkg/errors"
 )
