+++ ciao-controller/internal/datastore/sqlite3db_test.go
@@ -24,7 +24,7 @@ import (
 	"github.com/ciao-project/ciao/ciao-controller/types"
 	"github.com/ciao-project/ciao/ciao-storage"
 	"github.com/ciao-project/ciao/payloads"
-	"github.com/ciao-project/ciao/ssntp/uuid"
+	"github.com/ciao-project/ciao/uuid"
 )
 
 var dbCount = 1
