+++ ciao-controller/openstack_compute.go
@@ -23,7 +23,7 @@ import (
 	"github.com/ciao-project/ciao/ciao-controller/types"
 	"github.com/ciao-project/ciao/ciao-storage"
 	"github.com/ciao-project/ciao/openstack/compute"
-	"github.com/ciao-project/ciao/ssntp/uuid"
+	"github.com/ciao-project/ciao/uuid"
 	"github.com/gorilla/mux"
 )
