+++ ciao-deploy/cmd/auth-create.go
@@ -19,7 +19,7 @@ import (
 	"os"
 
 	"github.com/ciao-project/ciao/ciao-deploy/deploy"
-	"github.com/ciao-project/ciao/ssntp/uuid"
+	"github.com/ciao-project/ciao/uuid"
 
 	"github.com/spf13/cobra"
 )
