+++ ciao-scheduler/scheduler_ssntp_test.go
@@ -25,8 +25,8 @@ import (
 
 	"github.com/ciao-project/ciao/payloads"
 	"github.com/ciao-project/ciao/ssntp"
-	"github.com/ciao-project/ciao/ssntp/uuid"
 	"github.com/ciao-project/ciao/testutil"
+	"github.com/ciao-project/ciao/uuid"
 )
 
 /****************************************************************************/
