+++ networking/libsnnet/docker_plugin.go
@@ -29,7 +29,7 @@ import (
 	"time"
 
 	"github.com/ciao-project/ciao/database"
-	"github.com/ciao-project/ciao/ssntp/uuid"
+	"github.com/ciao-project/ciao/uuid"
 	"github.com/docker/go-connections/tlsconfig"
 	"github.com/docker/libnetwork/drivers/remote/api"
 	ipamapi "github.com/docker/libnetwork/ipams/remote/api"
