+++ payloads/assignpublicIP_test.go
@@ -20,8 +20,8 @@ import (
 	"testing"
 
 	. "github.com/ciao-project/ciao/payloads"
-	"github.com/ciao-project/ciao/ssntp/uuid"
 	"github.com/ciao-project/ciao/testutil"
+	"github.com/ciao-project/ciao/uuid"
 	"gopkg.in/yaml.v2"
 )
