+++ ssntp/frame.go
@@ -21,7 +21,7 @@ import (
 	"time"
 
 	"github.com/ciao-project/ciao/payloads"
-	"github.com/ciao-project/ciao/ssntp/uuid"
+	"github.com/ciao-project/ciao/uuid"
 )
 
 // TraceConfig is the SSNTP tracing configuration to be used
